requirements.txt

    selene>=2.0.0rc7
    selenium>=4.25.0
    pytest>=8.3.3
    allure-python-commons>=2.13.5
    allure-pytest